Detect if firewalld is running before trying to open port
As I am moving Gluster firewall to nftables, I need to manage
the firewall without firewalld for the time being, and detecting the
state of firewalld seems to be the cleanest way i found to avoid
a error. I could add another knob, but that's not a good design.

Thanks @Pilou_ for the condition simplification suggestion